[clang][docs] Improve "Obtainig Clang" section
The documentation is written relatively to `clang-llvm`, not the root repository directory. Therefore, the `llvm-project` repo needs to be cloned into the existing directory `clang-llvm`. As cloning into an existing directory is only allowed if the directory is empty, I added `mkdir ~/clang-llvm` to make the intent of creating an empty directory explicit.